agregar

+++ src/main/java/com/ejercicioclase/ejercicioclase/domain/repository/IdiomaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.ejercicioclase.ejercicioclase.domain.entities.Idioma;
 
-public interface IdiomaRepository extends CrudRepository<Idioma,Integer>{
+public interface IdiomaRepository extends CrudRepository<Idioma,Byte>{
 
 }
