+++ src/main/java/com/ejercicioclase/ejercicioclase/services/IdiomaService.java
@@ -6,7 +6,7 @@ import java.util.Optional;
 import com.ejercicioclase.ejercicioclase.domain.entities.Idioma;
 public interface IdiomaService {
      List<Idioma> findAll();
-Optional<Idioma> findById(Integer id);
+Optional<Idioma> findById(byte id);
 Idioma save(Idioma idioma);
 Idioma update(Long id);
 Optional<Idioma> delete(Long id);
